requirements.txt
Flask==2.3.2
gunicorn==21.2.0
tensorflow==2.20.0
numpy==1.24.3
pillow==10.0.0